# api/openapi.json
{
  "openapi": "3.0.0",
  "basePath": "/api",
  "info": {
    "title": "Toll connect",
    "version": "1.0.0",
    "description": "API for SoftEng 2024 NTUA course"
  },
  "definitions": {
    "TokenHeader": {
      "in": "header",
      "required": true,
      "name": "x-observatory-auth",
      "description": "authentication token",
      "schema": { "type": "string" }
    },
    "Format": {
      "in": "query",
      "required": false,
      "name": "format",
      "description": "format specification",
      "schema": { "type": "string", "enum": ["json", "csv"] }
    },
...
    "Error": {
      "type": "object",
      "properties": {
        "status": { "type": "string" },
        "info": { "type": "string" }
      },
      "required": ["status", "info"],
      "additionalProperties": true
    },
    "SuccessResponse": {
      "type": "object",
      "properties": {
        "status": { "type": "string", "description": "OK" },
        "info": { "type": "string", "description": "ok" }
      },
      "required": ["status", "info"]
    },
...
    "UnauthorizedResponse": {
      "description": "Unauthorized ",
      "content": {
        "application/json": { "schema": { "$ref": "#/definitions/Error" } }
      }
    },
    "BadRequestResponse": {
      "description": "Bad Request",
      "content": {
        "application/json": { "schema": { "$ref": "#/definitions/Error" } }
      }
    },
    "InternalServerErrorResponse": {
      "description": "Internal Server Error",
      "content": {
        "application/json": { "schema": { "$ref": "#/definitions/Error" } }
      }
    },
...
    "GetPaymentsResponse": {
      "type": "object",
      "properties": {
        "total_pages": {
          "type": "integer",
          "description": "Total number of pages"
        },
        "results": {
          "type": "array",
          "items": { "$ref": "#/definitions/Payment" }
        }
      },
      "required": ["total_pages", "results"]
    },
    "Payment": {
      "type": "object",
      "properties": {
        "_id": { "type": "string" },
        "payer": { "type": "string" },
        "payee": { "type": "string" },
        "dateofCharge": { "type": "string", "format": "date-time" },
        "amount": {
          "type": "number",
          "description": "The amount of the payment"
        },
        "dateofPayment": { "type": "string", "format": "date-time" },
        "dateofValidation": { "type": "string", "format": "date-time" },
        "__v": { "type": "integer", "description": "Version number (Mongoose)" }
      },
      "required": [
        "_id",
        "payer",
        "payee",
        "amount",
        "dateofCharge",
        "dateofPayment",
        "dateofValidation"
      ]
    },
...
    "TagsSchema": {
      "type": "object",
      "properties": {
        "_id": { "type": "string" },
        "tollOperator": {
          "type": "string",
          "description": "The Id of the Operator of the tag"
        },
        "__v": { "type": "integer", "description": "Version number (Mongoose)" }
      }
    },
    "TollsSchema": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "description": "Unique identifier for the toll"
        },
        "name": { "type": "string" },
        "latitude": { "type": "number", "format": "double" },
        "longitude": { "type": "number", "format": "double" },
        "locality": { "type": "string", "description": "The area of the toll" },
        "price": { "type": "array", "items": { "type": "number" } },
        "PM": { "type": "string" },
        "tollOperator": {
          "type": "string",
          "description": "The Id of the toll Operator"
        },
        "road": {
          "type": "string",
          "description": "The road where the toll is located"
        },
        "__v": { "type": "integer", "description": "Version number (Mongoose)" }
      },
      "required": [
        "_id",
        "name",
        "latitude",
        "longitude",
        "locality",
        "price",
        "PM",
        "tollOperator",
        "road",
        "__v"
      ]
    },
    "TollByOpResponse": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "description": "Unique identifier for the toll"
        },
        "name": { "type": "string" },
        "latitude": { "type": "number", "format": "double" },
        "longitude": { "type": "number", "format": "double" }
      },
      "required": ["_id", "name", "latitude", "longitude"]
    },
...
  },
  "paths": {
...
    "/api/payments/{status}/{date_from}/{date_to}": {
      "get": {
        "tags": ["Payments"],
        "summary": "Get payments",
        "operationId": "getPayments",
        "parameters": [
          {
            "name": "status",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["Created", "Paid", "Validated"]
            },
            "description": "Payment status"
          },
          {
            "name": "date_from",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "date" },
            "description": "Start date for filtering"
          },
          {
            "name": "date_to",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "date" },
            "description": "End date for filtering"
          },
          { "$ref": "#definitions/TokenHeader" },
          {
            "in": "query",
            "name": "page_size",
            "schema": { "type": "integer" },
            "description": "Number of results per page"
          },
          {
            "in": "query",
            "name": "page_number",
            "schema": { "type": "integer" },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "target_op_id",
            "schema": { "type": "string" },
            "description": "Target operator ID (optional)"
          },
          {
            "in": "query",
            "name": "is_payer",
            "schema": { "type": "boolean" },
            "description": "Filter by payer (optional)"
          },
          {
            "in": "query",
            "name": "is_payee",
            "schema": { "type": "boolean" },
            "description": "Filter by payee (optional)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of payments",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/definitions/GetPaymentsResponse" }
              }
            }
          },
          "400": { "$ref": "#/definitions/BadRequestResponse" },
          "401": { "$ref": "#/definitions/UnauthorizedResponse" },
          "500": { "$ref": "#/definitions/InternalServerErrorResponse" }
        }
      }
    },
    "/api/payments/pay/{id}": {
      "put": {
        "tags": ["Payments"],
        "summary": "Pay a payment",
        "operationId": "payPayment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID of the payment to be paid"
          },
          { "$ref": "#definitions/TokenHeader" }
        ],
        "responses": {
          "200": {
            "description": "Payment successfully paid",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/definitions/SuccessResponse" }
              }
            }
          },
          "400": { "$ref": "#/definitions/BadRequestResponse" },
          "401": { "$ref": "#/definitions/UnauthorizedResponse" },
          "500": { "$ref": "#/definitions/InternalServerErrorResponse" }
        }
      }
    },
...
    "/api/db/tolls/": {
      "get": {
        "tags": ["DB"],
        "summary": "Retrieve all toll documents",
        "operationId": "getAllTolls",
        "parameters": [
          { "$ref": "#/definitions/TokenHeader" },
          { "$ref": "#/definitions/Format" }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of toll documents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/definitions/TollsSchema" }
                }
              }
            }
          },
          "400": { "$ref": "#/definitions/BadRequestResponse" },
          "401": { "$ref": "#/definitions/UnauthorizedResponse" },
          "500": { "$ref": "#/definitions/InternalServerErrorResponse" }
        }
      }
    },
    "/api/db/tolls/{id}": {
      "get": {
        "tags": ["DB"],
        "summary": "Retrieve toll document with specified id",
        "operationId": "getTollWithID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Toll Id"
          },
          { "$ref": "#/definitions/TokenHeader" }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of payment document",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/definitions/TollsSchema" }
              }
            }
          },
          "400": { "$ref": "#/definitions/BadRequestResponse" },
          "401": { "$ref": "#/definitions/UnauthorizedResponse" },
          "500": { "$ref": "#/definitions/InternalServerErrorResponse" }
        }
      }
    },
    "/api/db/tolls/by_operator/{operator_id}": {
      "get": {
        "tags": ["DB"],
        "summary": "Retrieve toll documents belonging to specified operator",
        "operationId": "getTollsByOp",
        "parameters": [
          {
            "name": "operator_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Operator Id"
          },
          { "$ref": "#/definitions/TokenHeader" }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of toll documents",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/definitions/TollByOpResponse" }
              }
            }
          },
          "400": { "$ref": "#/definitions/BadRequestResponse" },
          "401": { "$ref": "#/definitions/UnauthorizedResponse" },
          "500": { "$ref": "#/definitions/InternalServerErrorResponse" }
        }
      }
    },
...
    "/api/db/payment/": {
      "get": {
        "tags": ["DB"],
        "summary": "Retrieve all payment documents",
        "operationId": "getAllPayments",
        "parameters": [
          { "$ref": "#/definitions/TokenHeader" },
          { "$ref": "#/definitions/Format" }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of payment documents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/definitions/Payment" }
                }
              }
            }
          },
          "400": { "$ref": "#/definitions/BadRequestResponse" },
          "401": { "$ref": "#/definitions/UnauthorizedResponse" },
          "500": { "$ref": "#/definitions/InternalServerErrorResponse" }
        }
      }
    },
    "/api/db/payment/{id}": {
      "get": {
        "tags": ["DB"],
        "summary": "Retrieve payment document with specified id",
        "operationId": "getPaymentWithID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Payment Id"
          },
          { "$ref": "#/definitions/TokenHeader" }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of payment document",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/definitions/Payment" }
              }
            }
          },
          "400": { "$ref": "#/definitions/BadRequestResponse" },
          "401": { "$ref": "#/definitions/UnauthorizedResponse" },
          "500": { "$ref": "#/definitions/InternalServerErrorResponse" }
        }
      }
    },
    "/api/db/tag/": {
      "get": {
        "tags": ["DB"],
        "summary": "Retrieve all tag documents",
        "operationId": "getAllTags",
        "parameters": [
          { "$ref": "#/definitions/TokenHeader" },
          { "$ref": "#/definitions/Format" }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of tag documents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/definitions/TagsSchema" }
                }
              }
            }
          },
          "400": { "$ref": "#/definitions/BadRequestResponse" },
          "401": { "$ref": "#/definitions/UnauthorizedResponse" },
          "500": { "$ref": "#/definitions/InternalServerErrorResponse" }
        }
      }
    },
    "/api/db/tag/{id}": {
      "get": {
        "tags": ["DB"],
        "summary": "Retrieve tag document with specified id",
        "operationId": "getTagWithID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Tag Id"
          },
          { "$ref": "#/definitions/TokenHeader" }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of tag document",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/definitions/TagsSchema" }
              }
            }
          },
          "400": { "$ref": "#/definitions/BadRequestResponse" },
          "401": { "$ref": "#/definitions/UnauthorizedResponse" },
          "500": { "$ref": "#/definitions/InternalServerErrorResponse" }
        }
      }
    }
  }
}